# static/api-docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "prose-pod-api",
		"description": "Prose Pod API server.",
		"contact": {
			"name": "Valerian Saliou",
			"email": "valerian@valeriansaliou.name"
		},
		"license": {
			"name": "MPL-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
...
		"/v1/server/config": {
			"put": {
				"tags": ["Init"],
				"summary": "Initialize the XMPP server",
				"description": "Initializes the XMPP server and returns its default configuration.",
				"operationId": "init_server_config",
				"security": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/InitServerConfigRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						},
						"headers": {
							"Location": { "$ref": "#/components/headers/Location" }
						}
					},
					"409": { "$ref": "#/components/responses/ServerConfigAlreadyInitialized" }
				}
			},
			"get": {
				"tags": ["Server / Configuration"],
				"summary": "Get server config",
				"description": "Get the current configuration of the XMPP server.",
				"operationId": "get_server_config",
				"security": [
					{ "BearerAuth": [] }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/allow-file-upload": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Change file upload settings",
				"description": "Activate or deactivate file upload and sharing.",
				"operationId": "set_file_upload_allowed",
				"security": [
					{ "BearerAuth": [] }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SetFileUploadingRequest" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/file-retention": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Change uploaded files retention",
				"description": "Change the retention of uploaded files.",
				"operationId": "set_file_retention",
				"security": [
					{ "BearerAuth": [] }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SetFileRetentionRequest" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/file-storage-encryption-scheme": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Change file storage encryption scheme",
				"description": "Change the file storage encryption scheme.",
				"operationId": "set_file_storage_encryption_scheme",
				"security": [
					{ "BearerAuth": [] }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/messaging/reset": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Reset messaging configuration",
				"description": "Reset the messaging configuration to its default value.",
				"operationId": "reset_messaging_config",
				"security": [
					{ "BearerAuth": [] }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/message-archive-retention": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Change message archive retention",
				"description": "Update message archive retention.",
				"operationId": "set_message_archive_retention",
				"security": [
					{ "BearerAuth": [] }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SetMessageArchiveRetentionRequest" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/message-archive-retention/reset": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Reset message archive retention",
				"description": "Reset the message archive retention to its default value.",
				"operationId": "reset_message_archive_retention",
				"security": [
					{ "BearerAuth": [] }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/server/config/store-message-archive": {
			"put": {
				"tags": ["Server / Configuration"],
				"summary": "Activate or deactivate message archiving",
				"description": "Activate or deactivate message archiving.",
				"operationId": "set_message_archive_enabled",
				"security": [
					{ "BearerAuth": [] }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SetMessageArchivingRequest" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"Error": {
				"type": "object",
				"required": [
					"reason"
				],
				"properties": {
					"reason": {
						"type": "string",
						"enum": [
							"bad_request",
							"database_error",
							"internal_server_error",
							"not_found",
							"not_implemented",
							"workspace_already_initialized",
							"workspace_not_initialized",
							"server_config_already_initialized",
							"server_config_not_initialized",
							"first_account_already_initialized",
							"unauthorized",
							"forbidden",
							"unknown"
						],
						"example": "bad_request"
					}
				}
			},
...
			"InitServerConfigRequest": {
				"type": "object",
				"required": [
					"domain"
				],
				"properties": {
					"domain": {
						"type": "string",
						"example": "crisp.chat"
					}
				}
			},
...
			"ServerConfig": {
				"type": "object",
				"required": [
					"domain",
					"message_archive_enabled",
					"message_archive_retention",
					"file_upload_allowed",
					"file_storage_encryption_scheme",
					"file_storage_retention",
					"mfa_required",
					"minimum_tls_version",
					"minimum_cipher_suite",
					"federation_enabled",
					"settings_backup_interval",
					"user_data_backup_interval"
				],
				"properties": {
					"domain": {
						"type": "string",
						"example": "crisp.chat"
					},
					"message_archive_enabled": {
						"type": "boolean",
						"default": true
					},
					"message_archive_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						],
						"default": "infinite"
					},
					"file_upload_allowed": {
						"type": "boolean",
						"default": true
					},
					"file_storage_encryption_scheme": {
						"type": "string",
						"default": "AES-256"
					},
					"file_storage_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						],
						"default": "infinite"
					},
					"mfa_required": {
						"title": "Require Multi-Factor Authentication",
						"description": "Whether or not [Multi-Factor Authentication (MFA)](https://en.wikipedia.org/wiki/Multi-factor_authentication) is required for users to log into their account.",
						"type": "boolean",
						"default": true
					},
					"minimum_tls_version": {
						"title": "Minimum SSL/TLS version",
						"description": "The minimum SSL/[TLS](https://fr.wikipedia.org/wiki/Transport_Layer_Security).",
						"type": "string",
						"default": "1.2"
					},
					"minimum_cipher_suite": {
						"title": "Minimum cipher suite (encryption strength)",
						"description": "Accepting lower-security encryption ciphers let older user devices and servers connect to your server, but is also much less secure. A minimum cipher suite of High strength is recommended.",
						"type": "string",
						"enum": ["HIGH_STRENGTH"],
						"default": "HIGH_STRENGTH"
					},
					"federation_enabled": {
						"description": "Whether or not the XMPP server should support interdomain federation. For more information, see [RFC 6120, section 13.14](https://www.rfc-editor.org/rfc/rfc6120#section-13.14).",
						"type": "boolean",
						"default": true
					},
					"settings_backup_interval": {
						"type": "string",
						"format": "duration",
						"default": "P1D"
					},
					"user_data_backup_interval": {
						"type": "string",
						"format": "duration",
						"default": "P1W"
					}
				}
			},
			"DurationDateLike": {
				"description": "A duration in [ISO 8601 format](https://en.wikipedia.org/wiki/ISO_8601#Durations) containing only date components.",
				"type": "string",
				"format": "duration",
				"example": "P1Y"
			},
...
			"DurationInfinite": {
				"type": "string",
				"enum": ["infinite"]
			},
...
			"SetFileRetentionRequest": {
				"type": "object",
				"required": [
					"file_retention"
				],
				"properties": {
					"file_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						]
					}
				}
			},
			"SetFileUploadingRequest": {
				"type": "object",
				"required": [
					"file_upload_allowed"
				],
				"properties": {
					"file_upload_allowed": {
						"type": "boolean"
					}
				}
			},
			"SetMessageArchiveRetentionRequest": {
				"type": "object",
				"required": [
					"message_archive_retention"
				],
				"properties": {
					"message_archive_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						]
					}
				}
			},
			"SetMessageArchivingRequest": {
				"type": "object",
				"required": [
					"message_archive_enabled"
				],
				"properties": {
					"message_archive_enabled": {
						"type": "boolean"
					}
				}
			},
...
		},
		"headers": {
			"Location": {
				"description": "URL to the created resource. See [`Location` on MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Location).",
				"schema": {
					"type": "string",
					"format": "uri"
				},
				"example": "/v1/invitations/42"
			},
			"WWW-Authenticate": {
				"description": "Standard `WWW-Authenticate` header. See [`WWW-Authenticate` on MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/WWW-Authenticate).",
				"schema": {
					"type": "string",
					"format": "www-authenticate"
				},
				"example": "Bearer realm=\"Admin only area\", charset=\"UTF-8\""
			},
...
		},
		"securitySchemes": {
			"BasicAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"responses": {
...
			"ServerConfigAlreadyInitialized": {
				"description": "XMPP server already initialized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "server_config_already_initialized"
						}
					}
				}
			},
			"ServerConfigNotInitialized": {
				"description": "XMPP server not initialized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "server_config_not_initialized"
						}
					}
				}
			},
...
			"Unauthorized": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "unauthorized"
						}
					}
				},
				"headers": {
					"WWW-Authenticate": { "$ref": "#/components/headers/WWW-Authenticate" }
				}
			},
			"Forbidden": {
				"description": "Forbidden",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "forbidden"
						}
					}
				}
			},
...
		},
...
	}
}